Remove gcry_fast_random_poll from _random_bytes
In Libgcrypt, gcry_fast_random_poll() is a macro that should be called
from time to time in order to add entropy to the internal pool of
the PRNG.  It is called from _random_add() after entropy is added to
the pool via gcry_random_add_bytes().

It is also called from _random_bytes(), but it seems unnecessary
to also call it immediately before obtaining random numbers.
As such, this commit removes the call to gcry_fast_random_poll()
from _random_bytes().

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>