Change buf parm type for random_bytes et al.
If a variable needs to be initialized to a random value, its address
must be cast to "unsigned char *" when calling random_bytes() or
random_pseudo_bytes().

This commit changes the "buf" parameter type from "unsigned char *"
to "void *" in the function prototypes for random_bytes(),
_random_bytes(), random_pseudo_bytes(), and _random_pseudo_bytes().
This eliminates the need for casting when calling these routines.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>