Add sharnanigans for testing executables
I call sharnanigans!!  (portmanteau of sharness shenanigans)

Add sharness tests for executables based on their manpage
documentation.  In particular, basic functionality is checked, munge
and unmunge are fully tested, and munged is a work-in-progress.

The shell function munged_stop_daemon() contains a busy-wait loop
for determining when the daemon has exited.  This is not ideal.

Tests requiring root privileges require the SUDO prerequisite which
is determined by checking whether non-interactive sudo is allowed.

Tested on:
- CentOS 6.9, 7.4.1708
- Debian 9.3
- Fedora 27
- FreeBSD 11.1
- NetBSD 7.1
- OpenBSD 6.2
- Ubuntu 16.04.3 LTS, 17.10

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>