Return error in munge/remunge clients for ttl < -1
The "--ttl" option in the munge and remunge clients is used to specify
a time-to-live value in seconds.

The clients and libmunge treat this value as a signed 32b integer
with all negative values being promoted to MUNGE_TTL_MAXIMUM (-1).

Internally, munged essentially treats this value as a uint32_t.
Its value is bounded by [-1, MUNGE_MAXIMUM_TTL]; the latter value is
set at compile-time in munge_defs.h to 3600.

This commit changes client behavior slightly.  The munge and remunge
clients now consider values < -1 or > UINT_MAX an error; note that
on 32b-systems, the maximum will be further limited to LONG_MAX.
This allows the clients to still specify the same 32b range as
before (which will be bounded by the daemon to MUNGE_MAXIMUM_TTL);
but anything < -1 is right out.  libmunge will pass this value on to
the daemon.  And the daemon will continue as before.

Tested on:
- CentOS 6.9, 7.4.1708
- Debian 9.3
- Fedora 27
- FreeBSD 11.1
- NetBSD 7.1
- OpenBSD 6.2
- Ubuntu 16.04.3 LTS, 17.10

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>