Add configure --with-munge-socket=PATH
The default path of the local domain socket (MUNGE_SOCKET_NAME) is
set relative to $(localstatedir) in "munge_defs.h".  Its value may
need to be modified when building a custom libmunge (for packaging,
perhaps) that communicates with the system-installed munged.

This commit adds the "--with-munge-socket=PATH" option to the configure
script to allow the default path of this socket to be overridden
without having to edit source files.

Tested on:
- Debian 9.3
- Fedora 27
- FreeBSD 11.1
- NetBSD 7.1
- OpenBSD 6.2

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>