Fix long test of 0000-sharness.t
The sharness test "EXPENSIVE prereq not activated by default"
fails when run with the "--long-tests" option.  When this option is
specified, the TEST_LONG environment variable is set and passed on to
the sub_test.  This causes the EXPENSIVE prereq to be met, which causes
sub_test 2 to not be skipped, which causes the overall test to fail.

Unset TEST_LONG from the environment to prevent passing this setting
to the sub_test.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>