Add munged "--stop" cmdline opt
Add the "-s/--stop" command-line option to munged.  This option
signals the PID of the process holding the exclusive advisory lock
for the specified socket.  It exits with a zero status if the daemon
was successfully stopped, or a non-zero status otherwise.

A SIGTERM is initially sent.  After a short (MUNGE_SIGNAL_DELAY_MSECS)
delay, a null signal is sent to check if the process is still running.
If so, additional SIGTERMs are sent using a linear back-off delay.
If the process fails to terminate, a SIGKILL is sent after
MUNGE_SIGNAL_ATTEMPTS.

This also adds the "-v/--verbose" command-line option.  It is currently
only used to issue a message indicating whether or not the daemon
has been successfully terminated (via SIGTERM) or killed (via SIGKILL).

Tested on:
- CentOS 6.9, 7.4.1708
- Debian 9.3
- Fedora 27
- FreeBSD 11.1
- NetBSD 7.1
- OpenBSD 6.2
- Ubuntu 16.04.3 LTS, 17.10

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>