Write pidfile after socket is successfully created
The PID file is created after the daemon has finished forking.
It should be written after any validation checks that might prevent
the daemon from starting, but before the original parent process
terminates.

It was discovered during testing that the PID file could be corrupted
if an attempt was made to start a new daemon using the same socket
name and PID file name as a daemon that was currently running.  The new
munged process would overwrite the PID file before attempting to obtain
the exclusive advisory lock protecting the socket.  Since this lock
was already being held by an active daemon, the lock attempt would
fail and the new munged process would terminate.  This would result
in the PID file now containing the PID of the just-terminated munged
process instead of the currently-running daemon.

This commit moves the creation of the PID file to after the advisory
lock has been obtained and the socket has been successfully created.

Tested on:
- CentOS 6.9, 7.4.1708
- Debian 9.3
- Fedora 27
- FreeBSD 11.1
- NetBSD 7.1
- OpenBSD 6.2
- Ubuntu 16.04.3 LTS, 17.10

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>