Replace "ln -s" with $(LN_S) in Makefiles
In Makefiles, replace the use of "ln -s" with $(LN_S).  This Makefile
variable is set by the AC_PROG_LN_S autoconf macro, and checks if
the operating system and file system support symbolic links.  $(LN_S)
is set to "ln -s" if symbolic links are supported, "ln" if only hard
links are supported, and "cp -pR" otherwise.

Note that AC_PROG_LN_S is not directly called from "configure.ac".
It appears to be called indirectly from another macro since the check
is already performed and $(LN_S) is set.

Reference:
- https://www.gnu.org/software/automake/manual/html_node/Extending.html
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.69/html_node/Particular-Programs.html

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>