gids: Add ghost_hash
Since the uid_hash is not persistent across gid_hash updates,
a non-existent user (i.e., one referenced in the group file, but
not listed in the passwd file) may be logged once for each gid_hash
construction cycle (repeated hourly by default).

This commit adds the ghost_hash to identify when a user first goes
missing from the passwd file.  This allows a given non-existent user to
be logged only once.  If this user is later added to the passwd file,
the next gid_hash update will clear the user from the ghost_hash,
thereby allowing the event to be re-logged should the user disappear
from the passwd file again.

Tested on:
- CentOS 6.9, 7.4.1708
- Debian 8.10, 9.3
- Fedora 27
- FreeBSD 11.1
- NetBSD 7.1
- OpenBSD 6.2
- Ubuntu 16.04.4 LTS, 17.10

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Issue #26