Fix HKDF mac_init assert failure for keylen > 0
The HKDF test vectors for case #3 and #6 in RFC 5869 specify a
salt length of 0 octets.  The HKDF salt is used as the HMAC key for
computing the PRK in the "extraction" phase.  In this first phase,
mac_init() starts by validating its parameters and assert()ing
'keylen' is > 0.  The salt length of 0 octets causes this assertion
to needlessly fail.

Update the 'keylen' assertion to ensure its value is >= 0.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>