Add MAINTAINERCLEANFILES
Add the MAINTAINERCLEANFILES automake variable to the top-level
"Makefile.am" for additional files to be cleaned via the
maintainer-clean target.

The files created by running the various autotools are added to
MAINTAINERCLEANFILES.  This contradicts the GNU Coding Standards,
but follows the advice from Russ Allbery's Build System Coding Style.
It allows the maintainer-clean target to reverse the result of running
the bootstrap script.

Reference:
- https://www.gnu.org/software/automake/manual/html_node/Clean.html
- https://www.gnu.org/prep/standards/html_node/Standard-Targets.html
- https://www.eyrie.org/~eagle/notes/style/build.html#automake

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>