Update configure.ac for backward compatibility
The dist-xz option is not recognized by automake-1.10.1 (2008,
Debian 5.0.10).  Use m4_ifdef to disable dist-xz on older versions
where the symbol is not defined.

The "tap-driver.sh" file is not installed by automake versions prior to
1.12 (e.g., automake-1.11.1 on CentOS 6.10, automake-1.11.3 on Ubuntu
12.04.5 LTS, automake-1.11.6 on Debian 7.11).  Copy "tap-driver.sh"
from automake-1.16.1 and store it in "build-aux/.tap-driver.sh.bak".
This file is copied to "build-aux/tap-driver.sh" by the bootstrap
script prior to running autoreconf.  This provides a working TAP
driver for older versions, while allowing automake-1.12 or later
to overwrite it (due to the --force-missing option).  Naming it
something besides "build-aux/tap-driver.sh" prevents changes to
it from being automatically tracked by git.  Naming it as a hidden
file (i.e., prefixing it with a ".") allows for "rm -f build-aux/*"
without removing the saved copy.

Specify the required version of autoconf in AC_PREREQ.  Testing
succeeds with autoconf-2.61 (2006, Debian 4.0) or later.

Specify the required version of automake in AM_INIT_AUTOMAKE along with
the automake options (moved from AUTOMAKE_OPTIONS in Makefile.am).
With dist-xz conditionally disabled and "tap-driver.sh" present,
testing succeeds with automake-1.10 (2006, Debian 4.0) or later.

Testing succeeds with libtool-2.2.6b (2008, CentOS 6.10 and Debian
6.0.10), but LT_INIT is not recognized by older versions (e.g.,
libtool-1.5.26 on Debian 5.0.10).  Use m4_ifdef to replace LT_INIT
(if not defined) with AC_PROG_LIBTOOL in order to support older
versions.  With AC_PROG_LIBTOOL, testing succeeds with libtool-1.5.22
(2005, Debian 4.0).  LT_PREREQ is not used since it appears to have
no effect on older versions; testing incorrectly succeeds with v1.5.26
and correctly fails with v2.2.6b when specifying a later version such
as LT_PREREQ([2.4.6]).

Reference:
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.69/html_node/Initializing-configure.html
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.69/html_node/Versioning.html
- https://www.gnu.org/software/automake/manual/html_node/Public-Macros.html
- https://www.gnu.org/software/automake/manual/html_node/API-Versioning.html
- https://www.gnu.org/software/libtool/manual/html_node/LT_005fINIT.html

Tested:
- CentOS 7.5.1804, 6.10
- Debian sid, 9.6, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora Rawhide, 29
- FreeBSD 11.2
- NetBSD 7.1.2
- OpenBSD 6.3
- Ubuntu 18.10, 18.04.1 LTS, 16.04.5 LTS, 14.04.5 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>