Fix undefined macro AM_PATH_LIBGCRYPT
The macro AM_PATH_LIBGCRYPT checks whether Libgcrypt exists on the
host system.  It is defined in "libgcrypt.m4" (which is typically
part of the Libgcrypt dev package).  If Libgcrypt is not installed,
the bootstrap script will fail with "error: possibly undefined macro:
AM_PATH_LIBGCRYPT".

Use m4_ifdef to disable AM_PATH_LIBGCRYPT when Libgcrypt is not
installed.

Reference:
- https://www.gnupg.org/documentation/manuals/gcrypt/Building-sources-using-Automake.html

Tested:
- Debian 8.11

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>