Fix assignment to readonly EUID variable in bash
In bash, the EUID variable expands to the effective user ID of the
current user, initialized at shell startup.  It is readonly.

In "t/0101-munged-security-socket.t", a sharness check uses a local
variable named EUID to store the results of $(id -u).  On systems
where /bin/sh links to bash, this results in the error "local: EUID:
readonly variable".

Fix the assignment to a potentially readonly variable by renaming EUID
to MY_EUID.  The EUID variable from bash is not being used since it
is not set by other shells such as dash.

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 9.7, 8.11, 7.11, 6.0.10, 5.0.10, 4.0, 3.1
- Fedora Rawhide, 29
- NetBSD 7.1.2
- OpenBSD 6.3
- Ubuntu 18.10, 18.04.1 LTS, 16.04.5 LTS, 14.04.5 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>