Fix expr warning wrt unportable BRE using '^'
In "t/0010-basic.t", a sharness check uses "^" as the first character
of the regex in an "expr STRING : REGEXP" test in order to anchor
the pattern.  On Debian 4 (but not later versions), this generates
the following warning:

> expr: warning: unportable BRE: `^MUNGE:.*:$': using `^' as the first
> character of the basic regular expression is not portable; it is being
> ignored

The "expr STRING : REGEXP" test is an anchored pattern match of REGEXP
in STRING.  As such, the leading "^" is superfluous.

Autoconf: Portable Shell Programming: Limitations of Usual Tools:
> Portable expr regular expressions should not begin with '^'.
> Patterns are automatically anchored so leading '^' is not needed
> anyway.

Remove the leading "^" in the regex.

Reference:
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.69/html_node/Limitations-of-Usual-Tools.html

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 9.7, 8.11, 7.11, 6.0.10, 5.0.10, 4.0, 3.1
- Fedora Rawhide, 29
- NetBSD 7.1.2
- OpenBSD 6.3
- Ubuntu 18.10, 18.04.1 LTS, 16.04.5 LTS, 14.04.5 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>