Set umask in munged_setup_env
munged_setup_env() sets up the environment for sharness checks.
This setup includes creating directories with appropriate permissions.
Directories are created via "mkdir -p" using the "-m" option to set
the directory mode.  While this mode is applied to the directory,
intermediate parent directories created via "-p" are permissioned
according to the current umask.

The Travis CI build now appears to be using a umask of 0.  This
results in intermediate directories created with 0777 permissions.
These permissions cause munged to fail with errors of "group-writable
permissions without sticky bit set".

Set the umask to 0022 to prevent creating directories with either
group-writable or world-writable permissions.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>