Add runstatedir
Modern Linux systems typically store runtime data in "/run" instead of
the traditional "/var/run".  This affects the location of the munged
socket name (MUNGE_SOCKET_NAME) and pidfile path (MUNGE_PIDFILE_PATH)
which are both stored in "${localstatedir}/run/munge".  Moving these
two files to "/run" is complicated by the fact that other files
(such as MUNGE_LOGFILE_PATH and MUNGE_SEEDFILE_PATH) are also stored
relative to ${localstatedir}, so simply setting this installation
directory variable is not a viable option.

Switch to ${runstatedir} for storing modifiable per-process data.
This variable defaults to "${localstatedir}/run", but can be set with
the "--runstatedir" configure option slated to appear in autoconf 2.70;
this option has also been backported to Debian's (and Ubuntu's)
autoconf 2.69-9.  If the configure script was built by an earlier
version of autoconf, ${runstatedir} can be set directly by appending
the variable assignment to the configure line.

Reference:
- https://lwn.net/Articles/436012/
- https://www.gnu.org/prep/standards/html_node/Directory-Variables.html
- https://lists.gnu.org/archive/html/bug-autoconf/2013-09/msg00002.html
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=759647

Tested:
- CentOS 7.6.1810 (autoconf 2.69-11), 6.10 (autoconf 2.63-5.1)
- Debian sid (autoconf 2.69-11), 9.8 (autoconf 2.69-10),
    8.11 (autoconf 2.69-8), 7.11 (autoconf 2.69-1),
    6.0.10 (autoconf 2.67-2), 5.0.10 (autoconf 2.61-8),
    4.0 (autoconf 2.61-4)
- Fedora 29 (autoconf 2.69-28)
- FreeBSD 12.0-RELEASE-p3 (autoconf 2.69_2)
- NetBSD 8.0 (autoconf 2.69nb8)
- OpenBSD 6.4 (autoconf 2.69p2)
- Ubuntu 18.10 (autoconf 2.69-11), 18.04.2 LTS (autoconf 2.69-11),
    16.04.6 LTS (autoconf 2.69-9), 14.04.6 LTS (autoconf 2.69-6),
    12.04.5 LTS (autoconf 2.68-1ubuntu2)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>