Fix use of dist-xz automake option
When configure.ac was updated for backward compatibility back in
commit b86188a, m4_ifdef was used to enable the "dist-xz" automake
option only if it was defined.  However, "dist-xz" is not a defined
macro, and as such, this option is never enabled.

Remove the m4_ifdef and always add "dist-xz" (and "no-dist-gzip") to
the options for AM_INIT_AUTOMAKE.  Additionally, update the minimum
version of automake to 1.11 which was the first release to support
xz compression.

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 9.9, 8.11, 7.11, 6.0.10
- Fedora 30
- FreeBSD 12.0-RELEASE-p3
- NetBSD 8.0
- OpenBSD 6.5
- Ubuntu 19.04, 18.04.2 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>