Allow decoding of 1-byte null origin address
For the origin address, the v3 credential format specifies an 8-bit
length (in bytes) followed by an IP address (in network byte order).
This allows the null address to be encoded in a single byte.  But the
origin address is currently always encoded as a 1-byte length followed
by a 4-byte IPv4 address.

Allow decoding of a null origin address that has been packed into a
single byte (where the address length is 0).  Encoding into a single
byte is being deferred until (presumably) the IPv6 work.

Tested:
- Debian 10.2

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>