Add munged "--origin" cmdline opt
Add the munged "--origin" command-line option to specify the hostname
or IP address for the origin address that will be encoded into
credential metadata.  This address must match an IP address assigned
to a local network interface.

The default value is the IPv4 address of the hostname returned by
gethostname().  If this does not match an IP address assigned to a
local network interface or if getifaddrs() is not found on the system,
munged will issue a warning message and use the null address for
its origin.  But if the "--origin" command-line option is specified,
failure to match an IP address will result in an error which can be
overridden via "--force".

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 9.9, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 30, 29
- FreeBSD 12.0-RELEASE-p5, 11.2-RELEASE-p10
- NetBSD 8.0, 7.1.2
- OpenBSD 6.5, 6.4
- openSUSE Leap 15.1
- Ubuntu 19.04, 18.04.2 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS

All systems tested had getifaddrs().  Consequently, HAVE_GETIFADDRS
was manually disabled in "config.h" to test the case where getifaddrs()
was not found on the system.

Also, the case where multiple IP addresses were assigned to the same
network interface was manually tested (on Debian sid).

Closes #69
Co-authored-by: Ward Poelmans <wpoely86@gmail.com>
Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>