Fix 0100-munged-lock check stopping munged as root
The sharness check 'stop unprivileged munged as root' in
"t/0100-munged-lock.t" tests whether root can stop a munged process
that was started by a non-privileged user.  If it fails, the munged
process continues running after the check finishes which hangs the
test suite.

Fix the sharness check so that if root fails to stop the non-privileged
process, a second attempt is made by the non-privileged user that
started it.

The construct "if foo; then :; else bar; fi" is used since negating
the exit status of commands with "!" is not portable.

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 10.0, 9.9, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 30, 29
- FreeBSD 12.0-RELEASE-p9, 11.2-RELEASE-p13
- NetBSD 8.0, 7.1.2
- OpenBSD 6.5, 6.4
- openSUSE Leap 15.1
- Ubuntu 19.04, 18.04.3 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>