Fix log msg for failure to query socket lockfile
While testing the case of having root stop a munged process started
by a non-privileged user, "munged --stop" returned an error with the
message "Lockfile not found".  However, the lockfile and socket both
existed, and the munged process was still running.  The source of
confusion was traced to the socket residing in an NFS mount exported
with root_squash, and the root process from "munged --stop" was
getting EACCES when attempting to open() the lockfile created with
S_IWUSR by a non-privileged user.

Update the log message text for a failure to query the socket lockfile
to better describe the cause of the underlying failure.  In particular,
differentiate between lock_query() returning 0 (lock not held) and
-1 (error occurred with errno set).  The above error now returns
"Permission denied".

Tested:
- Debian 9.9

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>