Disable valgrind sharnanigans in Travis CI config
The "t/0021-munged-valgrind.t" sharness test recently started failing
during Travis CI builds configured with libgcrypt.  The valgrind log
showed an unrecognized instruction:

> ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
> vex amd64->IR: unhandled instruction bytes: 0x48 0xF 0xC7 0xF0 0x72 0x4 0xFF 0xC9
> vex amd64->IR:   REX=1 REX.W=1 REX.R=0 REX.X=0 REX.B=0
> vex amd64->IR:   VEX=0 VEX.L=0 VEX.nVVVV=0x0 ESC=0F
> vex amd64->IR:   PFX.66=0 PFX.F2=0 PFX.F3=0
> valgrind: Unrecognised instruction at address 0x572cf23.
>    at 0x572CF23: ??? (in /lib/x86_64-linux-gnu/libgcrypt.so.20.0.5)
>    by 0x572CFD9: ??? (in /lib/x86_64-linux-gnu/libgcrypt.so.20.0.5)
>    by 0x572AE87: ??? (in /lib/x86_64-linux-gnu/libgcrypt.so.20.0.5)
>    by 0x572B811: ??? (in /lib/x86_64-linux-gnu/libgcrypt.so.20.0.5)
>    by 0x569A10A: ??? (in /lib/x86_64-linux-gnu/libgcrypt.so.20.0.5)
>    by 0x5696F98: gcry_control (in /lib/x86_64-linux-gnu/libgcrypt.so.20.0.5)
>    by 0x410D04: _random_add (random.c:582)
>    by 0x4100B9: _random_read_entropy

The Travis CI build is running in an Ubuntu 16.04.6 LTS image, but the
error could not be reproduced on an up-to-date Ubuntu 16.04.6 LTS VM.
Consequently, the sharness --long-tests (i.e., those tests marked with
the prereq EXPENSIVE which currently only include the valgrind tests)
are being disabled in the Travis CI config.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>