Fix dup err msg to stderr for failure to open logfile
When starting munged as root after having started it as a
non-privileged user, the following error message was written to
stderr twice:

> munged: Error: Logfile is insecure: "/var/log/munge/munged.log"
>   should be owned by UID 0 instead of UID 993
> munged: Error: Logfile is insecure: "/var/log/munge/munged.log"
>   should be owned by UID 0 instead of UID 993

When munged is started as a traditional daemon process (i.e., not
run in the foreground), log messages are initially written to stderr.
The process will double-fork() during daemonization and disassociate
from the tty.  Log messages will pivot to either the logfile or syslog.
If an error occurs before daemonization completes, the error message
will be relayed from the (grand)child process across the "daemonpipe"
to the parent process in order to be written to stderr.

In the error above, the first write was by the grandchild process
(i.e., after the double-fork() in daemonize_init()) before log
messages had pivoted to the logfile.  The grandchild process then
relayed the error message across the daemonpipe to have the parent
process write it to stderr before exiting.  This bug was introduced
in commit feccf43 when the daemonpipe abstraction was added.

Fix the duplicate message by having daemonpipe_write() only relay
the error message text if it has not already been written to stderr.

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 10.0, 9.9, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 30, 29
- FreeBSD 12.0-RELEASE-p9, 11.2-RELEASE-p13
- NetBSD 8.0, 7.1.2
- OpenBSD 6.5, 6.4
- openSUSE Leap 15.1
- Ubuntu 19.04, 18.04.3 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>