Do not reset log output stream in parent to relay err
In daemonize_init(), the parent process blocks on a read() from the
daemonpipe after forking in case an error message needs to be relayed
to stderr.  If an error message is received from the daemonpipe,
the log output stream is reset to stderr via log_open_file() and the
message is written via log_msg().  But resetting the output stream
can result in messages being logged with different options (e.g.,
LOG_OPT_TIMESTAMP).  Furthermore, it is unnecessary since the log
output stream remains set to stderr in the parent process.

Do not needlessly reset the log output stream in the parent process
when relaying an error message from the daemonpipe.

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 10.0, 9.9, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 30, 29
- FreeBSD 12.0-RELEASE-p9, 11.2-RELEASE-p13
- NetBSD 8.0, 7.1.2
- OpenBSD 6.5, 6.4
- openSUSE Leap 15.1
- Ubuntu 19.04, 18.04.3 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>