Disable parallel build for hkdf_test
Running "make check" in parallel occasionally causes the build to fail.
The error happens while building the libcommon, libmunge, and libtap
libraries needed for hkdf_test.  Presumably, a makefile dependency
is missing.

The best (and more portable) way to fix this is probably to switch
to non-recursive make.  But until then, the ".NOTPARALLEL" target
from GNU make can be used to force "hkdf_test" to be built serially.

Reference:
- https://www.gnu.org/software/make/manual/html_node/Special-Targets.html

Reproducer:
- while make check -j; do make clean -j; done

Tested:
- Debian 9.11

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>