Add $(AM_MAKEFLAGS) after $(MAKE)
When automake invokes make in a subdirectory, it uses the value of
the MAKE variable, and passes the value of the AM_MAKEFLAGS variable
to the make invocation.  However, AM_MAKEFLAGS is missing from the
makefile rule in "Make-inc.mk" responsible for rebuilding libraries
in their corresponding subdirectories.

Add $(AM_MAKEFLAGS) to the make invocation.

Reference:
- https://www.gnu.org/software/automake/manual/html_node/Subdirectories.html

Tested:
- Debian 9.11

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>