Add X_AC_WITH_RUNSTATEDIR
Add the autoconf macro X_AC_WITH_RUNSTATEDIR for specifying
"--with-runstatedir" in order to set "runstatedir".  This variable
defaults to "${localstatedir}/run".

This option is functionally equivalent to "--runstatedir" (slated
to appear in autoconf 2.70, and backported to Debian's and Ubuntu's
autoconf 2.69-9).  However, "--with-runstatedir" will override
"--runstatedir" if both are specified.

Since the GNU Coding Standards state that unrecognized options
must be accepted without halting, it seems safer to always enable
"--with-runstatedir" even if "--runstatedir" is available.

See commit 676b2ad for the initial support of runstatedir.

Reference:
- https://autotools.io/autoconf/arguments.html
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.69/html_node/External-Software.html

Tested:
- Debian 9.11

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>