Change munged "--origin" to allow any IPv4 address
Remove from the munged "--origin" command-line option the restriction
that the specified origin must match an IP address assigned to a
local network interface (see commit 5199aa4).  This removal allows a
configuration to specify the external address of a cluster that is
not assigned to a network interface on an internal node where the
credential is encoded.

Without this restriction, the origin address can be specified on
systems that do not support getifaddrs() as long as it is specified
by hostname or IP address (and not by interface name).

Ultimately, the security of the credential is determined by the key,
and the admin should be able to specify any origin address.  Using a
"false" origin will only affect other nodes within the MUNGE realm
sharing the same key, and it is the admin's determination whether
that use is appropriate.

Tested:
- Arch Linux
- CentOS Stream, 8.0.1905, 7.7.1908, 6.10
- Debian sid, 10.1, 9.11, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 31, 30
- FreeBSD 12.1-RELEASE-p1, 12.0-RELEASE-p12, 11.3-RELEASE-p5
- NetBSD 8.1, 7.2
- OpenBSD 6.6, 6.5
- openSUSE Leap 15.1
- Ubuntu 19.10, 18.04.3 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS

All systems tested had getifaddrs().  Consequently, HAVE_GETIFADDRS
was manually disabled in "config.h" to test the case where getifaddrs()
was not found on the system.

Issue #23
PR #69
Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>