Revisit "Logging stopped due to error" msg
The sharness test suite triggered several "Logging stopped due to
error" messages in /var/log/messages on FreeBSD, NetBSD, and OpenBSD.
These messages were not observed on any of the Linux test systems.

In log.c:_log_aux(), the formatted log message buffer is written to the
logfile via fprintf().  If this call fails, the aforementioned message
is logged via syslog() and further writes to the logfile are disabled.

The sharness test suite repeatedly starts and stops munged by sending
a SIGTERM (typically via "munged --stop").  On BSD, this SIGTERM
is occasionally interrupting fprintf() and causing it to fail with
errno=EINTR or with errno left unset.

Isolated tests of sending multiple SIGTERMs to a process calling
fprintf() in a loop showed 4KB of output (1 page) being dropped when
fprintf() returns an error.  Afterwards, writes to the output stream
resumed; consequently, disabling writes to the logfile (at least for
transient errors like EINTR) is unwarranted.

This commit makes the following changes:
1) The syslog() message includes the errno string from strerror()
   describing the error, and errno is initialized beforehand since
   it is not reliably set on an fprintf() error.
2) The syslog() message notes that messages may have been dropped.
3) The syslog() priority is reduced to LOG_ERR.
4) Further writes to the logfile are no longer disabled on error.

The syslog() message does not include the name of the logfile.
This is because the logging subsystem is given a FILE * and does
not have a record of the filename attached to the stream.  That may
be revisited in the future.  Also, it might make sense to disable
further writes to the logfile on persistent errors as long as this
behavior can be cleared on SIGHUP.

Tested:
- Debian 10.2
- FreeBSD 12.1-RELEASE-p1, 12.0-RELEASE-p12, 11.3-RELEASE-p5
- NetBSD 8.1, 7.2
- OpenBSD 6.6, 6.5

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>