Remove enable-arch & program-prefix from rpm %configure
Remove the "--enable-arch" and "--program-prefix" configure options
from the %configure line in "munge.spec".

The "--enable-arch" option was added in 2005 to select between building
RPMs for 32-bit/64-bit (commit e83a301).  But this option does not
appear to be useful for the general case and needlessly muddies the
spec file.

The "--program-prefix" option was added in 2003 to deal with
cross-compiling warnings (commit e0adeb1) and appears to be no longer
necessary.

Tested:
- CentOS Stream, 8.0.1905, 7.7.1908 (via rpmbuild)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>