Enable rpm _hardened_build
Set the RPM macro "_hardened_build" in "munge.spec" to ensure the
package is hardened.  This appears to be unnecessary since it is
the default for all packages beginning with Fedora 23.  However,
Fedora Packaging Guidelines state this must be enabled for packages
that are long running.

Setting _hardened_build adds "-fPIC" (if "-fPIE" is not already
present) to the compiler flags, and adds "-z now" to the linker
flags.  PIE adds security to executables by composing them entirely
of position-independent code.  Position-independent code (PIC)
is machine instruction code that executes properly regardless of
where in memory it resides.  PIE allows Exec Shield to use address
space layout randomization to prevent attackers from knowing where
existing executable code is during a security attack using exploits
that rely on knowing the offset of the executable code in the binary,
such as return-to-libc attacks.

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_pie
- https://fedoraproject.org/wiki/Changes/Harden_All_Packages

Tested:
- CentOS Stream, 8.0.1905, 7.7.1908 (via rpmbuild)
- Debian 10.2 (via make distcheck)
- Fedora 31, 30 (via fedpkg local)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>