Update rpm specfile
Remove the Group tags.  This tag has been deprecated since Fedora 17.
Also remove the "rm -rf %{buildroot}" in the first line of %install,
and the %clean section.

Update the License tags to better reflect the GPLv3+/LGPLv3+ nature.
The base package is now listed as GPLv3+, and the devel and libs
subpackages are listed as LGPLv3+.

Update the Requires tag in the base package to use a fully-versioned
arch-specific dependency since the executables within are linked
against libmunge in the libs subpackage.  The Fedora Packaging
Guidelines note:

  If you end up in a situation where the main package depends on the
  subpackage and the subpackage on the main package you should think
  carefully about why you don't have everything in the main package.

Perhaps libmunge should have been included in the base package instead
of a subpackage.  But having libmunge in a subpackage worked well
with multiarch since it meant you could install either a 32-bit or
64-bit munge package, and both 32-bit and 64-bit munge-libs packages.

Update the summary and descriptions for the devel and libs subpackages.

Move the %global _runstatedir macro definition from %prep to %build
in order to better document its use in the %configure line.

Add the "--with-crypto-lib=openssl" configure option.  This is
technically unnecessary since the configure script will check for
both OpenSSL and Libgcrypt.  But explicitly specifying the library
seems appropriate given the BuildRequires for openssl-devel.

Replace "make" with "%make_build", and "make install" with
"%make_install" (and not "%makeinstall" which should not be used).

Remove the "/" between %{buildroot} and a subsequent directory macro
since it results in a double-slash in the expanded text.

Change the directory permissions of %{_localstatedir}/lib/munge from
0711 to 0700.  The file-descriptor-passing authentication method is not
used on Linux, so the 0711 directory permissions are not needed here.

Remove the glob from shared libraries installed in %{_libdir}.
The glob conceals important parts of the file name, since changes to
the SONAME also typically result in a changed file name.  When the
library bumps its SONAME as part of an update, the change might remain
unnoticed and cause problems like broken dependencies.

Fedora Packaging Guidelines state that libtool archives should not be
included.  However, "libmunge.la" is being retained for now since its
removal should be treated as an API/ABI change.  It will be removed
on the next ABI change.  Note that Fedora does not include it in its
RPM package.

The %ldconfig_scriptlets macro is not used since it does not exist
on CentOS 7.

Packages were built on CentOS using rpmbuild:

  $ rpmbuild -tb munge-*.tar.xz

Packages were built on Fedora using fedpkg:

  $ fedpkg --release f31 local

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_tags_and_sections
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_requiring_base_package
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_parallel_make
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_why_the_makeinstall_macro_should_not_be_used
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_listing_shared_library_files
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#packaging-static-libraries
- https://docs.fedoraproject.org/en-US/packaging-guidelines/UsersAndGroups/
- https://docs.fedoraproject.org/en-US/quick-docs/creating-rpm-packages/

Tested:
- CentOS Stream, 8.0.1905, 7.7.1908 (via rpmbuild)
- Debian 10.2 (via make distcheck)
- Fedora 31, 30 (via fedpkg local)

On Fedora 31, the fedpkg local build reported a bad exit status:

  /usr/bin/rm -rf /home/dun/rpmbuild/BUILDROOT/munge-0.5.13-1.fc31.x86_64
  /usr/bin/rm: cannot remove '/home/dun/rpmbuild/BUILDROOT/munge-0.5.13-1.fc31.x86_64/usr/lib/debug/.dwz': Directory not empty
  error: Bad exit status from /var/tmp/rpm-tmp.pAEFuJ (%clean)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>