Replace rpm systemd Requires w/ %systemd_requires
Replace the 3 Requires directives for systemd with the
"%{?systemd_requires}" RPM macro in "munge.spec".  This macro is
present on recent versions of CentOS and Fedora:

  $ rpm --eval "%{?systemd_requires}"
  Requires(post): systemd
  Requires(preun): systemd
  Requires(postun): systemd

  $ rpm --showrc | grep -A4 systemd_requires
  -13: systemd_requires
  Requires(post): systemd
  Requires(preun): systemd
  Requires(postun): systemd
  %{nil}

The "BuildRequires: systemd" directive is being retained
instead of switching to "systemd-rpm-macros".  While Fedora has a
"systemd-rpm-macros" package, CentOS does not (as far as I can tell).
Specifying the "systemd" package instead seems to work on both.

Fedora Packaging Guidelines state that "systemd-rpm-macros" is
necessary for the %{_unitdir} macro to exist; however, this macro
appears to be present on both CentOS and Fedora with the inclusion
of the "systemd" package:

  $ rpm --eval %{_unitdir}
  /usr/lib/systemd/system

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/Scriptlets/#_systemd
- https://docs.fedoraproject.org/en-US/packaging-guidelines/Systemd/#_packaging

Tested:
- CentOS Stream, 8.0.1905, 7.7.1908 (via rpmbuild)
- Debian 10.2 (via make distcheck)
- Fedora 31, 30 (via fedpkg local)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>