Make sharness sudo-invoked commands EXPENSIVE
Make all sudo-invoked commands in the sharness test suite "expensive"
(i.e., requiring the EXPENSIVE prerequisite and enabled with the
sharness "--long-tests" option).  The rationale for this is that
these commands require sudo to be installed and configured to avoid
prompting the user for input of any kind.  That seems more suited to
a test environment instead of the general case.

TEST_LONG is checked instead of using "test_have_prereq EXPENSIVE"
since the EXPENSIVE prereq has not been set at this point by
"sharness.sh".

Tested:
- Arch Linux
- CentOS Stream, 8.0.1905, 7.7.1908, 6.10
- Debian sid, 10.2, 9.11, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 31, 30
- FreeBSD 12.1-RELEASE-p1, 11.3-RELEASE-p5
- NetBSD 8.1, 7.2
- OpenBSD 6.6, 6.5
- openSUSE Leap 15.1
- Ubuntu 19.10, 18.04.3 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>