Add logrotate config
Add a logrotate config file.  The location of the installed file can
be set via the "--with-logrotateddir" configure option.  It will
default to "${sysconfdir}/logrotate.d" if the "/etc/logrotate.d"
directory exists on the build system.

This fixes the following rpmlint warning:

  munge.x86_64: W: log-files-without-logrotate ['/var/log/munge']

munged does not (yet) close and reopen its log file upon receipt of
a SIGHUP.  Consequently, the logrotate config specifies "copytruncate"
in order to copy the log file to a new file, and truncate the original
log file in place after creating a copy.

For testing, the munge and munge-libs RPM packages were installed,
and munged was started:

  # rpm -ivh munge-0.5.13-2.fc31.x86_64.rpm \
      munge-libs-0.5.13-2.fc31.x86_64.rpm
  # sudo -u munge /usr/sbin/mungekey -v
  # systemctl start munge

The logrotate config was modified to adjust the "size" setting low
enough to trigger a rotation, and logrotate was invoked on the config:

  # sed -e 's/\<size .*/size 100/' /etc/logrotate.d/munge >/tmp/munge.$$
  # mv /tmp/munge.$$ /etc/logrotate.d/munge
  # ls -li /var/log/munge/munged.log
  # logrotate -v /etc/logrotate.d/munge
  # ls -li /var/log/munge/munged.log

"/var/log/munge/munged.log" was truncated to 0 bytes as expected.
The inode number remained the same after rotation.

A SIGHUP was sent to munged to trigger a log message:

  # kill -1 `cat /run/munge/munged.pid`

The log file then contained the new message indicting that munged
was still able to write to the log file after rotation.

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_logrotate_config_file

Tested:
- Fedora 31

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>